requirements.txt
pydantic==2.10.3
langchain==1.2.0
langchain-classic>=0.1.0
langchain-core>=0.3.0
langchain-groq==0.2.1
langchain-community==0.3.13
langchain-ollama>=0.2.0
langchain-huggingface==0.1.2
faiss-cpu==1.9.0
sentence-transformers==3.3.1
python-dotenv>=1.0.0
